moved s3.js to middleware folder and updated import
